% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jfurt.R, R/pizzar.R
\docType{package}
\name{pizzar}
\alias{pizzar}
\title{PIZZAR Package}
\usage{
pizzar(massa, quantidade)
}
\arguments{
\item{massa}{altura da massa: fina, media ou alta}

\item{quantidade}{quantas pizzas se quer fazer}
}
\description{
Retorna os ingredientes e suas quantidades para certo número de pizzas.
}
\details{
Dada uma quantidade de pizzas e seu tipo (massa fina, media ou alta) a funcão retorna os ingredientes e suas quantidades.
}
\examples{
pizzar("media", 3)

}
\author{
Mario Rutman /email{marrut@gmail.com}
}
